Redoing strike configuration to match scan
